+++ src/main/java/com/lxp/aplus/category/presentation/controller/CategoryController.java
@@ -1,6 +1,6 @@
 package com.lxp.aplus.category.presentation.controller;
 
-import com.lxp.aplus.category.application.dto.response.CategoryResult;
+import com.lxp.aplus.category.application.dto.result.CategoryResult;
 import com.lxp.aplus.category.application.port.in.CategoryQueryUseCase;
 import com.lxp.aplus.category.presentation.response.CategoryResponse;
 import com.lxp.aplus.common.result.ResultResponse;
